Create gene_parser.py
'''These functions can be used together to make lists of sequences for all genes
in a master file of blast outputs. The master file should be created using the
script makemasterDB.py, which concatenates a bunch of blast datasets into one
csv file. The file must be csv. The gene_parser.py script
must be executed from within a folder that contains the blastDatabase.csv file 
and all trinity files that correspond to that master blast file. As is, this 
program has to be executed from the python interpreter. If you rerun this file
it will replace any fasta files with the same output names.'''